MAINT: Switch to np.compat for long

+++ numba/core/types/__init__.py
@@ -105,8 +105,8 @@ int_ = _make_signed(np.int_)
 uint = _make_unsigned(np.int_)
 intc = _make_signed(np.intc) # C-compat int
 uintc = _make_unsigned(np.uintc) # C-compat uint
-long_ = _make_signed(np.long)
-ulong = _make_unsigned(np.long)
+long_ = _make_signed(int)  # int is an alias for long in Python3+
+ulong = _make_unsigned(int)
 longlong = _make_signed(np.longlong)
 ulonglong = _make_unsigned(np.longlong)
